eval-service/requirements.txt
# Foundry SDKs
azure-ai-projects>=2.0.0
azure-identity>=1.19.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# OpenTelemetry
opentelemetry-sdk>=1.29.0
azure-monitor-opentelemetry>=1.6.0
azure-core-tracing-opentelemetry>=1.0.0b12

# Cosmos DB
azure-cosmos>=4.9.0

# Utilities
pydantic>=2.10.0
httpx>=0.28.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
